added user validation in get,put blog api

+++ blog_backend/controllers/blogs.js
@@ -2,17 +2,27 @@ const router = require('express').Router()
 const Blog = require('../models/blog.js')
 
 router.get('/', async (request, response) => {
-    const result  = await Blog.find({}).populate('user',{username:1, name:1})
-    return response.status(200).json(result)
+    const user = request.user
+    if(user){
+        const result  = await Blog.find({user: user._id}).populate('user',{username:1, name:1})
+        return response.status(200).json(result)
+    }
+    return response.status(401).send('invalid token')
 })
 
 router.get('/:id', async (request, response) => {
-    const blog = await Blog.findById(request.params.id).populate('user',{username:1, name:1})
-    if(blog){
-        return response.status(200).json(blog)
+    const user = request.user
+    if(user && user.blogs.includes(request.params.id)){
+        const blog = await Blog.findById(request.params.id).populate('user',{username:1, name:1})
+        if(blog){
+            return response.status(200).json(blog)
+        }
+        else{
+            return response.status(404).send(`Blog with ${request.params.id} not found`)
+        }
     }
     else{
-        return response.status(404).send(`Blog with ${request.params.id} not found`)
+        return response.status(401).send('invalid token')
     }
 })
 
@@ -33,7 +43,7 @@ router.post('/', async (request, response) => {
 router.delete('/:id', async (request, response) => {
     const user = request.user
 
-    if(user.blogs.includes(request.params.id)){
+    if(user && user.blogs.includes(request.params.id)){
         const blog = await Blog.findByIdAndDelete(request.params.id)
         if(blog){
             //Updating user by deleting id in blogs array in User document.
@@ -53,18 +63,22 @@ router.delete('/:id', async (request, response) => {
 })
 
 router.put('/:id', async (request, response) => {
+    const user = request.user
 	const body = request.body
+    if(user && user.blogs.includes(request.params.id)){
+        const blogToUpdate = {
+            likes: body.likes
+        }
 
-	const blogToUpdate = {
-		likes: body.likes
-	}
-
-	const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blogToUpdate, { new : true, runValidators: true, context: 'query' })
-	if(updatedBlog){
-		response.json(updatedBlog)
-	}else{
-        return response.status(404).send(`Blog with ${request.params.id} not found`)
-	}
+        const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blogToUpdate, { new : true, runValidators: true, context: 'query' })
+        if(updatedBlog){
+            response.json(updatedBlog)
+        }else{
+            return response.status(404).send(`Blog with ${request.params.id} not found`)
+        }
+    }else{
+        return response.status(401).send('invalid token')
+    }
 })
 
 module.exports = router
